Use `long` instead of `integer` to prevent size overflows.

+++ x-pack/plugins/observability_solution/dataset_quality/server/services/data_telemetry/register_collector.ts
@@ -91,19 +91,19 @@ export function registerLogsDataUsageCollector(
             },
           },
           index_count: {
-            type: 'integer',
+            type: 'long',
             _meta: {
               description: 'Total number of indices in the stream of logs',
             },
           },
           namespace_count: {
-            type: 'integer',
+            type: 'long',
             _meta: {
               description: 'Total number of namespaces in the stream of logs',
             },
           },
           field_count: {
-            type: 'integer',
+            type: 'long',
             _meta: {
               description: 'Total number of fields in mappings of indices of the stream of logs',
             },
